with delete method and my hibernate dao

+++ src/main/java/com/evgeniy/spring/mvc/db/crud/dao/UserDaoImpl.java
@@ -1,15 +1,12 @@
 package com.evgeniy.spring.mvc.db.crud.dao;
 
 import com.evgeniy.spring.mvc.db.crud.models.User;
+import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
-import org.hibernate.query.Query;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
-import javax.persistence.criteria.CriteriaBuilder;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Root;
 import java.util.List;
 
 @Repository
@@ -20,32 +17,63 @@ public class UserDaoImpl implements DAO<User, Long>{
 
     @Override
     public void add(User entity) {
-        Session currentSession = sessionFactory.getCurrentSession();
-        currentSession.saveOrUpdate(entity);
+        try (Session session = sessionFactory.openSession()) {
+            session.beginTransaction();
+            session.save(entity);
+            session.getTransaction().commit();
+        } catch (HibernateException e) {
+            e.printStackTrace();
+        }
+    }
+
+    @Override
+    public void update(User entity) {
+        try (Session session = sessionFactory.openSession()) {
+            session.beginTransaction();
+            session.update(entity);
+            session.getTransaction().commit();
+        } catch (HibernateException e) {
+            e.printStackTrace();
+        }
     }
 
     @Override
     public User searchById(Long id) {
-        Session currentSession = sessionFactory.getCurrentSession();
-        User user = currentSession.get(User.class, id);
-        return user;
+        User user = null;
+        try (Session session = sessionFactory.openSession()) {
+            user = session.get(User.class, id);
+            if (user != null) {
+                return user;
+            } else {
+                System.out.println("User doesn't exist with provided Id..");
+            }
+        } catch (HibernateException e) {
+            e.printStackTrace();
+        }
+        return null;
     }
 
     @Override
     public void deleteById(Long id) {
-        Session session = sessionFactory.getCurrentSession();
-        User user = session.byId(User.class).load(id);
-        session.delete(user);
+        try (Session session = sessionFactory.openSession()) {
+            User user = searchById(id);
+            if (user != null) {
+                session.beginTransaction();
+                session.delete(user);
+                session.getTransaction().commit();
+            } else {
+                System.out.println("User doesn't exist with provided Id..");
+            }
+        } catch (HibernateException e) {
+            e.printStackTrace();
+        }
     }
 
     @Override
+    @SuppressWarnings("unchecked")
     public List<User> searchAll() {
-        Session session = sessionFactory.getCurrentSession();
-        CriteriaBuilder cb = session.getCriteriaBuilder();
-        CriteriaQuery<User> cq = cb.createQuery(User.class);
-        Root<User> root = cq.from(User.class);
-        cq.select(root);
-        Query query = session.createQuery(cq);
-        return query.getResultList();
+        try (Session session = sessionFactory.openSession()) {
+            return session.createQuery("from User", User.class).getResultList();
+        }
     }
 }
